Splitting sinatra/base.rb
