+++ lib/sinatra/errors.rb
@@ -1,0 +1,9 @@
+module Sinatra
+  class BadRequest < TypeError #:nodoc:
+    def http_status; 400 end
+  end
+
+  class NotFound < NameError #:nodoc:
+    def http_status; 404 end
+  end
+end
